+++ relatorios.h
@@ -1,0 +1,5 @@
+void tela_menu_relatorios();
+    void relat_clientes(void);
+    void relat_funcionarios(void);
+    void relat_produtos(void);
+    void relat_vendas(void);
