+++ vendas.h
@@ -1,0 +1,5 @@
+void tela_menu_vendas_pedidos();
+    void make_venda(void);
+    void exibir_pedido(void);
+    void edit_pedido(void);
+    void cancel_pedido(void);
